requirements.txt
pandas
SQLAlchemy
psycopg2-binary
chromadb
sentence-transformers
google-genai
python-dotenv